Make script run on all of GithHub
Why? The script only works if you directly land on the pull request page, and this triggers a document load. If you are on a repo's home page, then navigate to the pull request page it did not previously trigger.

There is also the possibility that you will get an "open new pull request button" on the home page of a site, and that should be patched as well.

I'd prefer a more targeted approach, maybe something like the "@include" TamperMonkey option?